Add files via upload
Extract all and open the Waddle-Chat-0.1.exe file inside the zip with the cmd (to have more control and also the function to change the port of the address by typing Waddle-Chat-0.1.exe -addr=":4097", of course the number can be whatever you want) or by clicking on it.
Then open your browser and type localhost:8080 (or the number you chose with the cmd) and you will be able to chat, if you have more PCs connected with the same IP, DNS and, obviously, Network, you can chat between them! It's fun, isn'it?